removeLivechatRoom
